Ability to skip components
We would also need a way to inject outputs from skipped components as those outputs should be needed by the subsequently processed components anyway.

For example, what if you skipped a docker-image component, but the subsequent terraform component needs the image repo and tag? You would need to provide the repo and the tag values and let the subsequent component work as if the image component was actually processed.

Thats another story though!