Add githubFiles driver
The githubFiles driver pushes changes and optionally runs kustomize-build to/on the target repo/branch via pull request.

This is usually used to update a GitOps repository with pre-requisites for subsequent steps in the workflow.

For example, you might want use this driver to add and update an ArgoCD application resource and Kubernetes manifests
along with the kustomization.yaml file to a GitOps repository.
The subsequent steps in the workflow would then deploy the application to the target cluster by updating the image tag
contained in the Kubernetes manifests pushed by this driver, which is then caught by the ArgoCD to deploy the application.